rag_support_project/rag_support_project/requirements.txt
langchain
langchain-community
langchain-google-genai
langgraph
chromadb
langchain-chroma
pypdf
python-dotenv
reportlab
streamlit
